REGRESSION (STP): Web Inspector search result loses selection in Elements tab rdar://153304495 https://bugs.webkit.org/show_bug.cgi?id=294396
Reviewed by Patrick Angle.

The bug can be simplified as, if you leave the Elements tab while
there's a node selected in the DOM tree, when you come back to the
Elements tab, that node appears unselected and can't be selected again
until you first select something else.

The regression point was https://github.com/WebKit/WebKit/pull/41834.
In that patch, we refactored and simplified the TreeOutline's logic for
forgetting an element and its descendants. During refactoring, we
removed the logic to ensure that those forgotten elements get
deselected. When the TreeOutline is set to visible the second time (when
you come back), the tree will try to reapply the selected element, but
it got skipped because its selectedTreeElement wasn't cleared. This
resulted in the selected style not being applied despite that element
is selected internally.

So, when all children of the TreeOutline get removed, we make sure to
clear the selectedTreeElement, restoring an equivalent behavior to the
code predating https://github.com/WebKit/WebKit/pull/41834. (Deselecting
all children by calling `deselect` still won't guarantee clearing
selectedTreeElement because the selected element could live deeper.)

Tested on macOS 26.0 and verified that the OpenResourceDialog still
works like normal.

* Source/WebInspectorUI/UserInterface/Views/TreeOutline.js:
(WI.TreeOutline.prototype._forgetTreeElementAndDescendants):

Canonical link: https://commits.webkit.org/296460@main