[text-wrap-style] should be applied to text in narrower containing blocks. https://bugs.webkit.org/show_bug.cgi?id=294777 <rdar://153942653>
Reviewed by Alan Baradlay.

`text-wrap-style` pretty is currently only being applied if the line width is wide enough
to fit 3 words. This causes an issue where resizing the containing block of a piece of text
can cause `text-wrap-style` to suddenly be applied/unapplied. We should instead apply
`text-wrap-style` as soon as the containing block is wide enough to fit the widest inline item.

This PR also fixes a test that was not correctly testing for the correct line breaks.

Combined changes:
* LayoutTests/fast/css3-text/css3-text-wrap/text-wrap-pretty-line-break-1-expected.html:
* LayoutTests/fast/css3-text/css3-text-wrap/text-wrap-style-narrow-block-expected.html: Added.
* LayoutTests/fast/css3-text/css3-text-wrap/text-wrap-style-narrow-block.html: Added.
* Source/WebCore/layout/formattingContexts/inline/InlineContentConstrainer.cpp:
(WebCore::Layout::validIdealLineWidth):

Canonical link: https://commits.webkit.org/296477@main