[Liquid Glass] [iOS] Smart App Banner is not blurred by scroll edge effect when top fixed elements are present https://bugs.webkit.org/show_bug.cgi?id=294805 rdar://151640309
Reviewed by Richard Robinson.

Currently, on iOS, the presence of a viewport-constrained element near one of the viewport edges
causes the scroll edge effect (i.e. scroll pocket) to be hidden, so that we get a solid color fill
in the inset area on that edge, instead of a blur / filter effect that blends the scroll view
background color. However, in the case where Safari shows a Smart App Banner or Tab Group banner (or
any other non-pinned banner UI), this policy causes the banner to not get blurred with the magic
pocket effect as the user scrolls, causing it to make text in the status bar illegible.

To fix this, we add logic to detect when the web view's scroll view is scrolled but the content
inset area above the top of the unobscured content rect is still visible, and force the magic pocket
to show in this case, even if there's a top fixed-position element.

* Source/WebKit/UIProcess/API/Cocoa/WKWebView.mm:
(-[WKWebView _updateFixedContainerEdges:]):
(-[WKWebView _updateTopScrollPocketCaptureColor]):

Note that we also need to explicitly override the capture color of the scroll pocket in the case
where there's a visible app banner in the content inset area. This is necessary because the capture
color defaults to the scroll view's background color, which may differ from the fixed element color
extension. For example, this avoids an issue on iPhone when visiting https://www.goodreads.com
where the fixed element color is orange, but the page background is white. If we don't update the
capture color to orange, we end up with a weird crossfade from orange to white in the magic pocket
under the smart app banner when scrolling.

(-[WKWebView _updateHiddenScrollPocketEdges]):
(-[WKWebView _addReasonToHideTopScrollPocket:]):
(-[WKWebView _removeReasonToHideTopScrollPocket:]):
(-[WKWebView _setTopScrollPocketHidden:]): Deleted.

Remove this separate method for hiding or showing the scroll pocket, and instead consolidate all the
logic into `-_updateHiddenScrollPocketEdges`, which is idempotent. This makes the logic a lot easier
to follow since we only need the `_shouldHideTopScrollPocket` check in one place, instead of having
to keep logic in `-_setTopScrollPocketHidden:` in sync with `-_updateHiddenScrollPocketEdges`.

* Source/WebKit/UIProcess/API/Cocoa/WKWebViewInternal.h:
* Source/WebKit/UIProcess/API/ios/WKWebViewIOS.h:
* Source/WebKit/UIProcess/API/ios/WKWebViewIOS.mm:
(-[WKWebView _updateNeedsTopScrollPocketDueToVisibleContentInset]):

See description above for more details.

(-[WKWebView _shouldHideTopScrollPocket]):

Consolidate the policy for determining whether to show the top scroll pocket into this helper
method. In short, `_reasonsToHideTopScrollPocket` (i.e. "are we in element fullscreen?") takes
precedence over `_needsTopScrollPocketDueToVisibleContentInset` ("is the app banner UI showing in
the content inset area?"), which in turn takes precedence over whether or not there's a top fixed-
position element color extension.

(-[WKWebView scrollViewDidChangeAdjustedContentInset:]):
(-[WKWebView scrollViewDidScroll:]):
* Source/WebKit/UIProcess/ios/WKContentView.mm:
(-[WKContentView didMoveToWindow]):

Additionally work around an existing Safari bug, rdar://153991882, which causes the scroll edge
effect to incorrectly disappear or appear when switching tabs. This is a necessary change alongside
the rest of the patch, because without it, there would be more cases where we end up with stale
hidden pocket edges state (causing the magic pocket to not appear when it should).

* Source/WebKit/UIProcess/mac/WebViewImpl.mm:
(WebKit::WebViewImpl::updateScrollPocket):

Canonical link: https://commits.webkit.org/296485@main