[WPE] WPE Platform: add initial API tests for wayland platform https://bugs.webkit.org/show_bug.cgi?id=294671
Reviewed by Carlos Alberto Lopez Perez.

Add an intial test to check we can create a WPEDisplayWayland and
connect to the wayland display. The script run-wpe-tests now detects if
a WPE Platform Wayland test is going to be run and it starts a weston
driver to ensure there's a wayland display to connect.

Canonical link: https://commits.webkit.org/296508@main