AX: accessibility/clip-path-bounding-box.html fails in macOS Tahoe because AccessibilityNodeObject and AccessibilityRenderObject don't explicitly enforce a minimum size for controls https://bugs.webkit.org/show_bug.cgi?id=294868 rdar://154135798
Reviewed by Joshua Hoffman.

With this commit, we now enforce a minimum size for controls for main-thread accessibility objects (matching the same
change done for AXIsolatedObjects some time ago), ensuring the rects we return are large enough for assistive technologies
to draw a cursor for. This is helpful for low-vision users who rely on magnifying the cursor contents — having no cursor
(as happens for 1x1px elements in ATs like VoiceOver) can be disorienting, especially for interactive elements like controls.

Fixes accessibility/clip-path-bounding-box.html for macOS Tahoe. We used to implicitly get this behavior, but no longer
do after the controls redesign in Tahoe, so this commit makes it explicit to fix the test. This logic, which was in
AXIsolatedObject, is now moved to the WebAccessibilityObjectWrapperMac so both object subclasses benefit from it.

* Source/WebCore/accessibility/AXCoreObject.h:
(WebCore::Accessibility::adjustControlRect):
* Source/WebCore/accessibility/isolatedtree/AXIsolatedObject.cpp:
(WebCore::AXIsolatedObject::relativeFrame const):

Canonical link: https://commits.webkit.org/296550@main