Slack is not recreating its web socket in case of network process crash rdar://153243346 https://bugs.webkit.org/show_bug.cgi?id=294842
Reviewed by Anne van Kesteren.

Slack is reacting to onclose but not onerror.
In case of network process crash, we were firing an error event but not a close event.
We now fire a close event, which gets us closer to spec.
We add a FAIL expectation to web-socket-tls test since, in debug, network process may crash, which will trigger a close event now.

* LayoutTests/http/tests/websocket/tests/hybi/network-process-crash-error-expected.txt:
* LayoutTests/http/tests/websocket/tests/hybi/network-process-crash-error.html:
* Source/WebKit/WebProcess/Network/WebSocketChannel.cpp:
(WebKit::WebSocketChannel::networkProcessCrashed):
* Tools/TestWebKitAPI/glib/TestExpectations.json:

Canonical link: https://commits.webkit.org/296607@main