Rework grid-self-baseline-two-dimensional test. https://bugs.webkit.org/show_bug.cgi?id=294943 rdar://problem/154246114
Reviewed by Alan Baradlay.

This test was imported from Blink as part of 206316@main. Since then,
there have been some discussions regarding how baseline alignment works
is a variety of scenarios such as orthogonal content and in particular
content that needs to synthesize a baseline. Blink ended up adopting
this behavior in:
https://chromium-review.googlesource.com/c/chromium/src/+/3916247

As part of that change they also updated a variety of WPTs to reflect
this new behavior and also deleted this test from their repo. In this
patch I do two main things:

1.) Move this test over to WPT. For the most part this is just moving
the test to the correct location and linking to the correct WPT
infrastructure rather than the copies we had locally. I also adjusted
the font sizes to follow the recommendations in:

https://web-platform-tests.org/writing-tests/ahem.html

which is just using the font shorthand and making sure the size is a
multiple of 5px.

2.) Update the expectations to reflect this new behavior.

Canonical link: https://commits.webkit.org/296619@main