Allow self to be passed as validation anchor https://bugs.webkit.org/show_bug.cgi?id=294903
Reviewed by Ryosuke Niwa.

In 292770@main we made the control the fallback when the validation
anchor was not given. This created a slight inconsistency with the API
which would throw if you passed the control as validation anchor.

This changes the API to allow passing the control as per step 8 of
https://html.spec.whatwg.org/multipage/custom-elements.html#dom-elementinternals-setvalidity
(Note that the specification implements the fallback as part of the API
call, but that is not how we implement it.)

Test changes are already upstreamed:

https://github.com/web-platform-tests/wpt/pull/53353
Canonical link: https://commits.webkit.org/296622@main