[visionOS] Gaze glow shape is incorrect for elements with associated label and non-uniform border radii or clip path https://bugs.webkit.org/show_bug.cgi?id=294952 rdar://154258426
Reviewed by Abrar Rahman Protyasha.

During the shrink wrapping of interaction regions, we previously
collected the rect of each discovered region, and then set the
clip path equal to the result of `pathWithShrinkWrappedRects`
with the collected rects passed as a parameter.

Now, for each discovered region, we only add its rect to the
collection if the region does not have a clip path. If a clip
path is present, we add it a collection of clip paths. The
final clip path is then set to a path consisting of all of the
discovered clip paths + the result of `pathWithShrinkWrappedRect`
for the collected rects.

* LayoutTests/interaction-region/clip-path-with-label-expected.txt: Added.
* LayoutTests/interaction-region/clip-path-with-label.html: Added.
* Source/WebCore/rendering/EventRegion.cpp:
(WebCore::EventRegionContext::shrinkWrapInteractionRegions):

Canonical link: https://commits.webkit.org/296646@main