[Skia] Symbols with the emoji variation selector are rendered with the text version https://bugs.webkit.org/show_bug.cgi?id=294966
Reviewed by Adrian Perez de Castro.

The system fallback fonts code only supports querying the font by a
single charater, so in case of a cluster consisting of base character +
emoji variation selector, the variation selector is ignored and if
there's a non-emoji font supporting the base character it's used instead.
In that case we can override the base character with an emoji character
when querying the font to try to ensure an emoji font is retrieved.

Canonical link: https://commits.webkit.org/296658@main