ssue caused by vertical-align on root node with rem/rex type https://bugs.webkit.org/show_bug.cgi?id=294451 rdar://151618114
Reviewed by Sam Weinig.

Issue caused by vertical-align on root node with rem/rex type
https://bugs.webkit.org/show_bug.cgi?id=294451
rdar://151618114

Reviewed by Sam Weinig.

The issue is when doing vertical-align of type rem or rex, it is assumed that there
is a valid root style. This is not so in case of root node and this was causing
issues (root style is NULL ofr root element).
The fix was built over the following PR: https://github.com/WebKit/WebKit/pull/46917.
The usage of new BuilderConverter::convertLength as opposed to builderState.cssToLengthConversionData in function BuilderCustom::applyValueVerticalAlign fixed the issue combined with the earlier
changes fixed the issue

Canonical link: https://commits.webkit.org/296678@main