AX: string for text marker range should not skip an emitted newline if it's the first character https://bugs.webkit.org/show_bug.cgi?id=295017 rdar://154368379
Reviewed by Tyler Wilcock.

The user-facing issue was that VoiceOver would fail to speak the text
of a new line in a document when navigating by line using caret
navigation, because VoiceOver thought the new cursor position was
visually equivalent to the previous one.

The underlying cause was that if a text marker range consisted of only
an emitted newline, the newline was being suppressed by mistake, so it
made it seem like the range was empty, indicating that two positions
were equivalent when in fact they differ due to the emitted newline.

* LayoutTests/accessibility/isolated-tree/simple-word-navigation-expected.txt:
* LayoutTests/accessibility/mac/text-marker-emitted-newlines-expected.txt: Added.
* LayoutTests/accessibility/mac/text-marker-emitted-newlines.html: Added.
* Source/WebCore/accessibility/AXTextMarker.cpp:
(WebCore::AXTextMarkerRange::toString const):

Canonical link: https://commits.webkit.org/296690@main