REGRESSION(283753@main): Video as image source with `decoding="async"` is decoded always synchronously https://bugs.webkit.org/show_bug.cgi?id=295066 rdar://153915118
Reviewed by Simon Fraser.

The bug is in RenderBoxModelObject::decodingModeForImageDraw() when we choose the
DecodingMode for the image and the element has the attribute decoding="async”.
In this case we try to be smart and override the user choice if it is going to
cause flickering. We check if the element is in the viewport and if it is, we do
use DecodingMode::Synchronous.

But this should logic not be applied to animated images (or videos as images)
because for animated images we calculate the decoding mode for the next frame.

* Source/WebCore/platform/graphics/BitmapImage.cpp:
(WebCore::BitmapImage::draw):
* Source/WebCore/rendering/RenderBoxModelObject.cpp:
(WebCore::RenderBoxModelObject::decodingModeForImageDraw const):

Canonical link: https://commits.webkit.org/296706@main