functionGenerateHeapSnapshot should expect an OOM exception in JSONParse rdar://150692273 https://bugs.webkit.org/show_bug.cgi?id=294991
Reviewed by Keith Miller.

The crash reported in the radar is an assertion failure that may happen in
`functionGenerateHeapSnapshot` when it calls `JSONParse` to parse a json string. The call
is followed by an assertion that no exception was thrown. However, in low memory
conditions parsing may legitimately throw an OutOfMemory exception. This happens in ASAN
builds when they run the test case reported in the radar. In non-ASAN builds OOM is not
thrown when running the same test case.

The patch changes the shapshot function to expect and propagate exceptions thrown in the
parse call.

* Source/JavaScriptCore/jsc.cpp:
(JSC_DEFINE_HOST_FUNCTION):

Canonical link: https://commits.webkit.org/296739@main