[Grid] Fix grid-container-ignores-first-letter-001 https://bugs.webkit.org/show_bug.cgi?id=295114 rdar://154504582
Reviewed by Alan Baradlay.

The first-letter text comes from the "first formatted line of its
originating element." Finding the first formatted line of a block
container:
  - Comes from the inline level content of the first line box if the
    block container establishes an inline formatting context.
  - If it contains block-level content, then it comes from the first
    formatted line of the first in-flow block-level child. If no such
    line exists, then there is no formatted line.

The flexbox and grid specs state that they do not contribute a first
formatted line or letter.

In theory, finding the first formatted line requires performing layout,
but we construct first letter renderers and associated structures during
render tree building/updating, which occurs before.
RenderBlock::firstLetterAndContainer attempts to implement a heuristic
in which it finds the block container that has the first letter style
and then digs into its content to find the renderer that would be
associated with that style. During this process, if we hit a flexbox or a
grid, we seem to just assign firstLetter to the next sibling and continue
the heuristic. Instead, we should *not* return a first letter container
and renderer since there should be no first formatted line and, as a
result, no first letter.

* LayoutTests/css3/flexbox/flexbox-ignore-container-firstLetter-expected.txt:
* LayoutTests/css3/flexbox/flexbox-ignore-container-firstLetter.html:
* LayoutTests/fast/css-grid-layout/grid-container-ignore-first-letter.html: Removed.
This test seems to be a duplicate of the one contained in the grid WPT
directory (grid-container-ignores-first-letter-001).

* LayoutTests/imported/w3c/web-platform-tests/css/css-flexbox/flexbox-ignores-first-letter.html
This is the same as the grid test but for flexbox.
* LayoutTests/imported/w3c/web-platform-tests/css/css-flexbox/flexbox-ignores-first-letter-expected.txt: Added.
* LayoutTests/imported/w3c/web-platform-tests/css/css-flexbox/flexbox-ignores-first-letter.html: Added.
* LayoutTests/imported/w3c/web-platform-tests/css/css-grid/grid-model/grid-container-ignores-first-letter-001-expected.txt:
* Source/WebCore/rendering/RenderBlock.cpp:

* LayoutTests/css3/flexbox/flexbox-ignore-container-firstLetter-expected.txt
* LayoutTests/css3/flexbox/flexbox-ignore-container-firstLetter.html
Flexboxes to not generate a first formatted line or a first letter for
their ancestors. As a result, the container does not have a first
formatted line/letter since the first in flow child does not.

Canonical link: https://commits.webkit.org/296797@main